Finish draft of transfer learning example
